# Improve Fused8BitRowwiseQuantizedSBFloatToFloatOrHalfNeon by 5%-15%